Refactor: Extract RippleEffect from Moire.tsx
