+++ src/components/effects/Moire/Moire.test.tsx
@@ -1,4 +1,3 @@
-
 import { render, waitFor } from "@testing-library/react";
 
 // Mock chroma-js before importing the component
@@ -10,12 +9,12 @@ jest.mock("chroma-js", () => ({
 // We need to avoid using 'document' in the factory, so we'll use a placeholder
 // and swap it out or mock the appendChild method to accept our placeholder.
 const mockCanvas = {
-    nodeName: 'CANVAS',
-    style: {},
-    parentNode: null,
-    width: 100,
-    height: 100,
-    // Add a fake nodeType to mimic a Node if needed, but appendChild checks strict type
+  nodeName: "CANVAS",
+  style: {},
+  parentNode: null,
+  width: 100,
+  height: 100,
+  // Add a fake nodeType to mimic a Node if needed, but appendChild checks strict type
 };
 
 jest.mock("ogl", () => {
@@ -25,40 +24,45 @@ jest.mock("ogl", () => {
       setSize;
       render;
       constructor() {
-          this.gl = {
-            canvas: mockCanvas,
-            clearColor: jest.fn(),
-            renderer: {
-                extensions: { OES_texture_half_float: {} },
-                isWebgl2: false,
-                render: jest.fn() // Added render function here
-            },
-            HALF_FLOAT: 1,
-            FLOAT: 2,
-            RGBA32F: 3,
-            RGBA16F: 4,
-            RGBA: 5,
-          };
-          this.setSize = jest.fn();
-          this.render = jest.fn();
+        this.gl = {
+          canvas: mockCanvas,
+          clearColor: jest.fn(),
+          renderer: {
+            extensions: { OES_texture_half_float: {} },
+            isWebgl2: false,
+            render: jest.fn(), // Added render function here
+          },
+          HALF_FLOAT: 1,
+          FLOAT: 2,
+          RGBA32F: 3,
+          RGBA16F: 4,
+          RGBA: 5,
+        };
+        this.setSize = jest.fn();
+        this.render = jest.fn();
       }
     },
     Camera: class {
-        position = { set: jest.fn(), z: 50 };
-        perspective = jest.fn();
-        fov = 45;
-        aspect = 1;
+      position = { set: jest.fn(), z: 50 };
+      perspective = jest.fn();
+      fov = 45;
+      aspect = 1;
     },
     Geometry: class {},
     Program: class {
-        uniforms = { tDiffuse: { value: null }, uCenter: { value: { set: jest.fn() } }, uRadius: { value: 0 }, uStrength: { value: 0 } };
+      uniforms = {
+        tDiffuse: { value: null },
+        uCenter: { value: { set: jest.fn() } },
+        uRadius: { value: 0 },
+        uStrength: { value: 0 },
+      };
     },
     Mesh: class {},
     Color: class {
-        set = jest.fn();
+      set = jest.fn();
     },
     Vec2: class {
-        set = jest.fn();
+      set = jest.fn();
     },
     RenderTarget: class {},
     Triangle: class {},
@@ -79,14 +83,18 @@ describe("MagicComponent (Moire)", () => {
     });
 
     // Mock HTMLDivElement.prototype.appendChild to allow our fake canvas
-    jest.spyOn(HTMLDivElement.prototype, 'appendChild').mockImplementation((node) => {
+    jest
+      .spyOn(HTMLDivElement.prototype, "appendChild")
+      .mockImplementation((node) => {
         // Just ignore the type check failure simulation or return the node
         return node as Node;
-    });
+      });
 
-    jest.spyOn(HTMLDivElement.prototype, 'removeChild').mockImplementation((node) => {
+    jest
+      .spyOn(HTMLDivElement.prototype, "removeChild")
+      .mockImplementation((node) => {
         return node as Node;
-    });
+      });
   });
 
   afterAll(() => {
@@ -97,14 +105,21 @@ describe("MagicComponent (Moire)", () => {
     const addEventListenerSpy = jest.spyOn(window, "addEventListener");
     const removeEventListenerSpy = jest.spyOn(window, "removeEventListener");
     const _docAddEventListenerSpy = jest.spyOn(document, "addEventListener");
-    const docRemoveEventListenerSpy = jest.spyOn(document, "removeEventListener");
+    const docRemoveEventListenerSpy = jest.spyOn(
+      document,
+      "removeEventListener",
+    );
 
     const { unmount } = render(<MagicComponent />);
 
     // Wait for the setTimeout(..., 0) to initialize the component
     await waitFor(() => {
-        // We expect resize listener to be added
-        expect(addEventListenerSpy).toHaveBeenCalledWith("resize", expect.any(Function), false);
+      // We expect resize listener to be added
+      expect(addEventListenerSpy).toHaveBeenCalledWith(
+        "resize",
+        expect.any(Function),
+        false,
+      );
     });
 
     // Unmount the component
@@ -112,9 +127,17 @@ describe("MagicComponent (Moire)", () => {
 
     // Verify cleanup
     // We expect removeEventListener to be called for resize
-    expect(removeEventListenerSpy).toHaveBeenCalledWith("resize", expect.any(Function), false);
+    expect(removeEventListenerSpy).toHaveBeenCalledWith(
+      "resize",
+      expect.any(Function),
+      false,
+    );
 
     // Check if scroll listener is removed
-    expect(docRemoveEventListenerSpy).toHaveBeenCalledWith("scroll", expect.any(Function), { passive: true });
+    expect(docRemoveEventListenerSpy).toHaveBeenCalledWith(
+      "scroll",
+      expect.any(Function),
+      { passive: true },
+    );
   });
 });
